Generic worker: move artifact feature code into correct file
