+++ workers/generic-worker/artifacts.go
@@ -4,15 +4,11 @@ import (
 	"encoding/json"
 	"fmt"
 	"log"
-	"mime"
 	"net"
 	"net/url"
-	"os"
-	"path/filepath"
 	"runtime"
 	"strconv"
 	"strings"
-	"time"
 
 	"github.com/taskcluster/httpbackoff/v3"
 	tcurls "github.com/taskcluster/taskcluster-lib-urls"
@@ -31,210 +27,6 @@ var (
 		".log": "text/plain",
 	}
 )
-
-// PayloadArtifacts returns the artifacts as listed in the payload of the task (note this does
-// not include log files)
-func (task *TaskRun) PayloadArtifacts() []artifacts.TaskArtifact {
-	payloadArtifacts := make([]artifacts.TaskArtifact, 0)
-	for _, artifact := range task.Payload.Artifacts {
-		basePath := artifact.Path
-		base := &artifacts.BaseArtifact{
-			Name:     artifact.Name,
-			Expires:  artifact.Expires,
-			Optional: artifact.Optional,
-		}
-		// if no name given, use canonical path
-		if base.Name == "" {
-			base.Name = canonicalPath(basePath)
-		}
-		// default expiry should be task expiry
-		if time.Time(base.Expires).IsZero() {
-			base.Expires = task.Definition.Expires
-		}
-		switch artifact.Type {
-		case "file":
-			payloadArtifacts = append(payloadArtifacts, resolve(base, "file", basePath, artifact.ContentType, artifact.ContentEncoding, task.pd))
-		case "directory":
-			if errArtifact := resolve(base, "directory", basePath, artifact.ContentType, artifact.ContentEncoding, task.pd); errArtifact != nil {
-				payloadArtifacts = append(payloadArtifacts, errArtifact)
-				continue
-			}
-			walkFn := func(path string, d os.DirEntry, incomingErr error) error {
-				subPath, err := filepath.Rel(taskContext.TaskDir, path)
-				if err != nil {
-					// this indicates a bug in the code
-					panic(err)
-				}
-				relativePath, err := filepath.Rel(basePath, subPath)
-				if err != nil {
-					// this indicates a bug in the code
-					panic(err)
-				}
-				subName := filepath.Join(base.Name, relativePath)
-				b := &artifacts.BaseArtifact{
-					Name:     canonicalPath(subName),
-					Expires:  base.Expires,
-					Optional: base.Optional,
-				}
-				switch {
-				// Issue 6488
-				// Sometimes an error occurs during the scanning of a file or
-				// directory. Perhaps this is when a file/directory is deleted
-				// while the directory is being scanned, or perhaps it can
-				// happen if Generic Worker does not have read permissions for
-				// the file/directory. Either way, create an error artifact to
-				// cause the task to fail, and the cause to be preserved in the
-				// error artifact.
-				case incomingErr != nil:
-					fullPath := filepath.Join(taskContext.TaskDir, subPath)
-					payloadArtifacts = append(
-						payloadArtifacts,
-						&artifacts.ErrorArtifact{
-							BaseArtifact: b,
-							Message:      fmt.Sprintf("Error processing file '%s' as artifact: %s", fullPath, incomingErr),
-							Reason:       "invalid-resource-on-worker",
-							Path:         subPath,
-						},
-					)
-				case d.IsDir():
-					if errArtifact := resolve(b, "directory", subPath, artifact.ContentType, artifact.ContentEncoding, task.pd); errArtifact != nil {
-						payloadArtifacts = append(payloadArtifacts, errArtifact)
-					}
-				default:
-					payloadArtifacts = append(payloadArtifacts, resolve(b, "file", subPath, artifact.ContentType, artifact.ContentEncoding, task.pd))
-				}
-				return nil
-			}
-			// Any error returned here should already have been handled by
-			// walkFn, so should be safe to ignore.
-			_ = filepath.WalkDir(filepath.Join(taskContext.TaskDir, basePath), walkFn)
-		}
-	}
-	return payloadArtifacts
-}
-
-// File should be resolved as an S3Artifact if file exists as file and is
-// readable, otherwise i) if it does not exist as a "file-missing-on-worker" ErrorArtifact,
-// or ii) if it cannot be read by the task user, as a "file-not-readable-on-worker" ErrorArtifact,
-// otherwise if it exists as a directory, as an "invalid-resource-on-worker" ErrorArtifact.
-// A directory should resolve as `nil` if directory exists as directory and is readable,
-// otherwise i) if it does not exist or ii) cannot be read, as a "file-missing-on-worker"
-// ErrorArtifact, otherwise if it exists as a file, as
-// "invalid-resource-on-worker" ErrorArtifact
-// TODO: need to also handle "too-large-file-on-worker"
-func resolve(base *artifacts.BaseArtifact, artifactType, path, contentType, contentEncoding string, pd *process.PlatformData) artifacts.TaskArtifact {
-	fullPath := filepath.Join(taskContext.TaskDir, path)
-	fileReader, err := os.Open(fullPath)
-	if err != nil {
-		// cannot read file/dir, create an error artifact
-		return &artifacts.ErrorArtifact{
-			BaseArtifact: base,
-			Message:      fmt.Sprintf("Could not read %s '%s'", artifactType, fullPath),
-			Reason:       "file-missing-on-worker",
-			Path:         path,
-		}
-	}
-	defer fileReader.Close()
-	// ok it exists, but is it right type?
-	fileinfo, err := fileReader.Stat()
-	if err != nil {
-		return &artifacts.ErrorArtifact{
-			BaseArtifact: base,
-			Message:      fmt.Sprintf("Could not stat %s '%s'", artifactType, fullPath),
-			Reason:       "invalid-resource-on-worker",
-			Path:         path,
-		}
-	}
-	if artifactType == "file" && fileinfo.IsDir() {
-		return &artifacts.ErrorArtifact{
-			BaseArtifact: base,
-			Message:      fmt.Sprintf("File artifact '%s' exists as a directory, not a file, on the worker", fullPath),
-			Reason:       "invalid-resource-on-worker",
-			Path:         path,
-		}
-	}
-	if artifactType == "directory" && !fileinfo.IsDir() {
-		return &artifacts.ErrorArtifact{
-			BaseArtifact: base,
-			Message:      fmt.Sprintf("Directory artifact '%s' exists as a file, not a directory, on the worker", fullPath),
-			Reason:       "invalid-resource-on-worker",
-			Path:         path,
-		}
-	}
-	if artifactType == "directory" {
-		return nil
-	}
-
-	tempPath, err := copyToTempFileAsTaskUser(fullPath, pd)
-	if err != nil {
-		return &artifacts.ErrorArtifact{
-			BaseArtifact: base,
-			Message:      fmt.Sprintf("Could not copy file '%s' to temporary location as task user: %v", fullPath, err),
-			Reason:       "file-not-readable-on-worker",
-			Path:         path,
-		}
-	}
-
-	// Is content type specified in task payload?
-	if contentType == "" {
-		extension := filepath.Ext(path)
-		// first look up our own custom mime type mappings
-		contentType = customMimeMappings[strings.ToLower(extension)]
-		// then fall back to system mime type mappings
-		if contentType == "" {
-			contentType = mime.TypeByExtension(extension)
-		}
-		// lastly, fall back to application/octet-stream in the absense of any other value
-		if contentType == "" {
-			// application/octet-stream is the mime type for "unknown"
-			contentType = "application/octet-stream"
-		}
-	}
-	// Is content encoding specified in task payload?
-	if contentEncoding == "" {
-		extension := filepath.Ext(path)
-		// originally based on https://github.com/evansd/whitenoise/blob/03f6ea846394e01cbfe0c730141b81eb8dd6e88a/whitenoise/compress.py#L21-L29
-		SkipCompressionExtensions := map[string]bool{
-			".7z":    true,
-			".bz2":   true,
-			".deb":   true,
-			".dmg":   true,
-			".flv":   true,
-			".gif":   true,
-			".gz":    true,
-			".jpeg":  true,
-			".jpg":   true,
-			".npz":   true,
-			".png":   true,
-			".swf":   true,
-			".tbz":   true,
-			".tgz":   true,
-			".webp":  true,
-			".whl":   true, // Python wheel are already zip file
-			".woff":  true,
-			".woff2": true,
-			".xz":    true,
-			".zip":   true,
-			".zst":   true,
-		}
-		// When the file extension is blacklisted in SkipCompressionExtensions then "identity" should be used, otherwise "gzip".
-		if SkipCompressionExtensions[extension] {
-			contentEncoding = "identity"
-		} else {
-			contentEncoding = "gzip"
-		}
-	}
-	return createDataArtifact(base, fullPath, tempPath, contentType, contentEncoding)
-}
-
-// The Queue expects paths to use a forward slash, so let's make sure we have a
-// way to generate a path in this format
-func canonicalPath(path string) string {
-	if os.PathSeparator == '/' {
-		return path
-	}
-	return strings.ReplaceAll(path, string(os.PathSeparator), "/")
-}
 
 // createDataArtifact creates a TaskArtifact for the given data, according to
 // the CreateObjectArtifacts configuration.
